+++ tests/test_async.py
@@ -1,30 +1,32 @@
 import os
+import unittest
 
 import aiohttp
-import asynctest
 import brawlstats
 import pytest
+
 from dotenv import load_dotenv
 
 pytestmark = pytest.mark.asyncio
 load_dotenv()
 
 
-class TestAsyncClient(asynctest.TestCase):
-    use_default_loop = True
+class TestAsyncClient(unittest.IsolatedAsyncioTestCase):
 
     PLAYER_TAG = '#V2LQY9UY'
     CLUB_TAG = '#UL0GCC8'
 
-    async def setUp(self):
-        session = aiohttp.ClientSession(loop=self.loop)
-
+    async def asyncSetUp(self):
+        session = aiohttp.ClientSession(trust_env=True)
         self.client = brawlstats.Client(
-            os.getenv('token'),
-            base_url=os.getenv('base_url'),
-            is_async=True,
-            session=session
+            token=os.getenv('TOKEN'),
+            session=session,
+            base_url=os.getenv('BASE_URL'),
+            is_async=True
         )
+
+    async def asyncTearDown(self):
+        await self.client.close()
 
     async def test_get_player(self):
         player = await self.client.get_player(self.PLAYER_TAG)
@@ -34,6 +36,9 @@ class TestAsyncClient(asynctest.TestCase):
         club = await player.get_club()
         self.assertIsInstance(club, brawlstats.Club)
         self.assertEqual(club.tag, self.CLUB_TAG)
+
+        battle_logs = await player.get_battle_logs()
+        self.assertIsInstance(battle_logs, brawlstats.BattleLog)
 
         with self.assertRaises(brawlstats.NotFoundError):
             await self.client.get_player('2PPPPPPP')
@@ -71,7 +76,8 @@ class TestAsyncClient(asynctest.TestCase):
         self.assertIsInstance(club_members, brawlstats.Members)
         self.assertIn(self.PLAYER_TAG, [x.tag for x in club_members])
 
-        await self.assertAsyncRaises(brawlstats.NotFoundError, self.client.get_club_members('8GGGGGGG'))
+        with self.assertRaises(brawlstats.NotFoundError):
+            await self.client.get_club_members('8GGGGGGG')
 
     async def test_get_rankings(self):
         player_ranking = await self.client.get_rankings(ranking='players')
@@ -104,22 +110,14 @@ class TestAsyncClient(asynctest.TestCase):
         with self.assertRaises(ValueError):
             await self.client.get_rankings(ranking='brawlers', brawler='SharpBit')
 
-    async def test_get_constants(self):
-        constants = await self.client.get_constants()
-        self.assertIsInstance(constants, brawlstats.Constants)
-
-        maps = await self.client.get_constants('maps')
-        self.assertIsInstance(maps, brawlstats.Constants)
-
-        await self.assertAsyncRaises(KeyError, self.client.get_constants('invalid'))
-
     async def test_get_brawlers(self):
         brawlers = await self.client.get_brawlers()
         self.assertIsInstance(brawlers, brawlstats.Brawlers)
 
-    async def asyncTearDown(self):
-        await self.client.close()
+    async def test_get_event_rotation(self):
+        events = await self.client.get_event_rotation()
+        self.assertIsInstance(events, brawlstats.EventRotation)
 
 
 if __name__ == '__main__':
-    asynctest.main()
+    unittest.main()