+++ tests/test_blocking.py
@@ -13,11 +13,13 @@ class TestBlockingClient(unittest.TestCase):
     CLUB_TAG = '#UL0GCC8'
 
     def setUp(self):
-
         self.client = brawlstats.Client(
-            os.getenv('token'),
-            base_url=os.getenv('base_url')
+            token=os.getenv('TOKEN'),
+            base_url=os.getenv('BASE_URL')
         )
+
+    def tearDown(self):
+        self.client.close()
 
     def test_get_player(self):
         player = self.client.get_player(self.PLAYER_TAG)
@@ -27,6 +29,9 @@ class TestBlockingClient(unittest.TestCase):
         club = player.get_club()
         self.assertIsInstance(club, brawlstats.Club)
         self.assertEqual(club.tag, self.CLUB_TAG)
+
+        battle_logs = player.get_battle_logs()
+        self.assertIsInstance(battle_logs, brawlstats.BattleLog)
 
         self.assertRaises(brawlstats.NotFoundError, self.client.get_player, '2PPPPPPP')
         self.assertRaises(brawlstats.NotFoundError, self.client.get_player, 'P')
@@ -82,21 +87,13 @@ class TestBlockingClient(unittest.TestCase):
         self.assertIsInstance(us_brawler_ranking, brawlstats.Ranking)
         self.assertTrue(len(us_brawler_ranking) == 1)
 
-    def test_get_constants(self):
-        constants = self.client.get_constants()
-        self.assertIsInstance(constants, brawlstats.Constants)
-
-        maps = self.client.get_constants('maps')
-        self.assertIsInstance(maps, brawlstats.Constants)
-
-        self.assertRaises(KeyError, self.client.get_constants, 'invalid')
-
     def test_get_brawlers(self):
         brawlers = self.client.get_brawlers()
         self.assertIsInstance(brawlers, brawlstats.Brawlers)
 
-    def tearDown(self):
-        self.client.close()
+    def test_get_event_rotation(self):
+        events = self.client.get_event_rotation()
+        self.assertIsInstance(events, brawlstats.EventRotation)
 
 
 if __name__ == '__main__':